requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
av==16.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
CTkTable==1.1
customtkinter==5.2.2
daphne==4.2.1
darkdetect==0.8.0
Django==5.2.9
exceptiongroup==1.3.0
fastapi==0.118.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.30.0
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
mysql-connector-python==9.4.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
Pillow
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyOpenSSL==25.3.0
pyparsing==3.2.3
python-decouple==3.8
python-vlc==3.0.21203
pywin32-ctypes==0.2.3

requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.48.0

tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
zope.interface==8.1.1

redis==5.2.1